Update the setup.py template to match xstatic standard
The content of setup.py for the xstatic packages is standard, and it has
been arrived at with a lot of pain and suffering.  All of its parts
serve a purpose in one corner case or other, and any changes should be
made very carefully and with testing.

In this case, particularly important are the "version" field, which is
needed for the installed packages to have proper egg-info metadata, and
the "namespace_packages" field, which makes sure the package is
installed in the proper namespace.

It's also important to not add arbitrary things to the "setup_requires"
field, as many build systems don't have unrestricted access to the
internet and can't just download the additional dependencies on demand.